scraper/requirements.txt
httpx>=0.27.0
supabase>=2.0.0
pydantic>=2.8.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
google-genai>=1.0.0
tqdm>=4.66.0
